+++ CHANGELOG.md
@@ -15,6 +15,7 @@
   * Cubbhole Secrets Engine: Add support for read secret, read paths, write secret and delete secret APIs.
   * Database Secrets Engine: Add support for generating dynamic DB credentials.
   * GoogleCloud Secrets Engine: Add support for generating OAuth2 Token & Service Account Key.
+  * Nomad Secrets Engine: Add support for generating dynamic credentials.
 
 ## 0.10.4001 (August 10, 2018)